requirements.txt
mmh3
pycrypto